+++ src/tasks/TaskWindow.cpp
@@ -79,7 +79,7 @@ void TaskWindow::NotifyProgress(Task * task)
 
 void TaskWindow::OnTick()
 {
-	intermediatePos = std::fmod(ui::Engine::Ref().LastTick() * 0.06f, 100.f);
+	intermediatePos = float(std::fmod(ui::Engine::Ref().LastTick() * 0.06, 100.0));
 	task->Poll();
 	if (done)
 		Exit();
